Move the vendored-runtime support code under tasks/shared/vendor
